# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/onboarding": {
      "get": {
        "operationId": "AdminOnboardingController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingChecklistResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-onboarding"
        ]
      }
    },
    "/api/v1/admin/onboarding/{code}/complete": {
      "post": {
        "operationId": "AdminOnboardingController_complete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingChecklistResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-onboarding"
        ]
      }
    },
    "/api/v1/admin/onboarding/{code}/skip": {
      "post": {
        "operationId": "AdminOnboardingController_skip",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingChecklistResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-onboarding"
        ]
      }
    },
    "/api/v1/admin/store/publish": {
      "post": {
        "operationId": "AdminPublicationController_publish",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingChecklistResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-onboarding"
        ]
      }
    },
    "/api/v1/admin/store/unpublish": {
      "post": {
        "operationId": "AdminPublicationController_unpublish",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingChecklistResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-onboarding"
        ]
      }
    },
...
    "/api/v1/admin/catalog/import-template": {
      "get": {
        "operationId": "AdminCatalogImportController_template",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/api/v1/admin/catalog/import": {
      "post": {
        "operationId": "AdminCatalogImportController_import",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogImportSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "ciadelivery",
    "description": "ciadelivery HTTP API",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OnboardingStepResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "create_tenant",
              "create_store",
              "configure_branding",
              "configure_domain",
              "configure_address",
              "configure_hours",
              "configure_delivery",
              "create_owner",
              "import_catalog",
              "connect_whatsapp",
              "create_couriers",
              "place_test_order",
              "validate_notifications",
              "train_team",
              "publish_store"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "DONE",
              "SKIPPED"
            ]
          },
          "doneAt": {
            "type": "string",
            "nullable": true
          },
          "doneBy": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "code",
          "status",
          "doneAt",
          "doneBy",
          "note"
        ]
      },
      "OnboardingChecklistResponse": {
        "type": "object",
        "properties": {
          "published": {
            "type": "boolean"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OnboardingStepResponse"
            }
          },
          "nextCode": {
            "type": "string",
            "nullable": true,
            "enum": [
              "create_tenant",
              "create_store",
              "configure_branding",
              "configure_domain",
              "configure_address",
              "configure_hours",
              "configure_delivery",
              "create_owner",
              "import_catalog",
              "connect_whatsapp",
              "create_couriers",
              "place_test_order",
              "validate_notifications",
              "train_team",
              "publish_store"
            ]
          },
          "blockingCodes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "create_tenant",
                "create_store",
                "configure_branding",
                "configure_domain",
                "configure_address",
                "configure_hours",
                "configure_delivery",
                "create_owner",
                "import_catalog",
                "connect_whatsapp",
                "create_couriers",
                "place_test_order",
                "validate_notifications",
                "train_team",
                "publish_store"
              ]
            }
          }
        },
        "required": [
          "published",
          "steps",
          "nextCode",
          "blockingCodes"
        ]
      },
      "OnboardingNoteDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CatalogImportSummaryResponse": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "number"
          },
          "categories": {
            "type": "number"
          },
          "products": {
            "type": "number"
          },
          "optionGroups": {
            "type": "number"
          },
          "options": {
            "type": "number"
          }
        },
        "required": [
          "rows",
          "categories",
          "products",
          "optionGroups",
          "options"
        ]
      },
...
    }
  }
}